Make CacheService return nil when read/write/delete are called if there is no cache available.

+++ lib/caches/mc_cache.rb
@@ -7,24 +7,40 @@ class McCache
     end
 
     def initialize
-      config = ConfigService.load_config('memcached.yml')[ConfigService.environment]
-      @@client ||= Dalli::Client.new(config['host'], config)
-    rescue => error
-      SdkLogger.logger.error("McCache.initialize error: #{error.message}")
-      @@client ||= Dalli::Client.new('localhost:11211', { host: 'localhost:11211',namespace: 'mc_cache', compress: true })
+      @@config ||= ConfigService.load_config('memcached.yml')[ConfigService.environment]
+      @@client ||= Dalli::Client.new(@@config['host'], @@config)
+      cache_exists?
+    rescue Exception => error
+      puts("McCache.initialize error: #{error.message}")
+      @@client = nil
     end #initialize
+
+    def cache_exists?
+    #   @@client.read('test')
+    # rescue Exception => error
+    #   @@client = nil
+      Net::HTTP.get(URI("http://#{@@config['host']}"))
+    rescue Errno::ECONNREFUSED => error
+      puts "**** Error: #{error.message}"
+      @@client = nil
+    rescue Net::HTTPBadResponse => error
+      # do nothing
+    end
 
     # Cache API, mimics ActiveSupport::Cache::Store
     # http://api.rubyonrails.org/classes/ActiveSupport/Cache/Store.html
     def read(key, options = {})
+      return unless client
       client.get(key, options)
     end
 
     def write(key, value, options = {})
+      return unless client
       client.set(key, value, options[:expires_in], options)
     end
 
     def delete(key, options = {})
+      return unless client
       deleted = read(key)
       client.delete(key)
       deleted
